refactor: update base interfaces and re-name files
