+++ src/app/(private)/order/page.tsx
@@ -19,7 +19,7 @@ import {
   OrderProjectsTotal,
 } from "./_components/order-projects";
 import { OrderReferenceForm } from "./_components/order-reference-form";
-import { OrderSummaryTable } from "./_components/order-summary-table";
+import { SummaryTable } from "./_components/summary-table";
 
 export default async function OrderPage() {
   return (
@@ -48,7 +48,7 @@ export default async function OrderPage() {
           <FormContent>
             <OrderProjects>
               <OrderProjectsContent>
-                <OrderSummaryTable />
+                <SummaryTable />
                 <OrderProjectsActions />
               </OrderProjectsContent>
               <OrderProjectsTotal />
