+++ src/types/piece.d.ts
@@ -1,0 +1,6 @@
+interface Piece {
+  id: string;
+  name?: string;
+  qtde: number;
+  material: Material;
+}
