+++ src/types/project.d.ts
@@ -1,0 +1,18 @@
+type Cost = {
+  name: string;
+  qtde: number;
+  value: number;
+};
+
+type Project = {
+  name: string;
+  qtde: number;
+
+  pieces: Piece[];
+
+  costs: Cost[];
+
+  profitRate: number;
+  monthlyExpense: number;
+  comission: number;
+};
